[Silver III] Title: 주유소, Time: 100 ms, Memory: 5648 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -2,9 +2,9 @@
 #define ll long long
 using namespace std;
 
-ll solution(int n, vector<int> road, vector<int> city) {
-    int sum = 0;
-    int sel = 1e9;
+ll solution(int n, vector<ll> road, vector<ll> city) {
+    ll sum = 0;
+    ll sel = 1e9;
     for (int i = 0; i < n - 1; i++) {
         sel = min(sel, city[i]);
         sum += sel * road[i];
@@ -14,18 +14,18 @@ ll solution(int n, vector<int> road, vector<int> city) {
 
 int main() {
     int N;
-    vector<int> road;
-    vector<int> city;
+    vector<ll> road;
+    vector<ll> city;
 
     cin >> N;
     for (int i = 0; i < N - 1; i++) {
-        int input;
+        ll input;
         cin >> input;
         road.push_back(input);
     }
 
     for (int i = 0; i < N; i++) {
-        int input;
+        ll input;
         cin >> input;
         city.push_back(input);
     }
